+++ handlers/admin/information_section.go
@@ -1,0 +1,6 @@
+package admin
+
+const (
+	// InformationSectionWeight controls the order of the Information page in admin navigation.
+	InformationSectionWeight = 170
+)
