--- handlers/information/section.go
@@ -1,6 +1,0 @@
-package information
-
-const (
-	// SectionWeight controls the order of the Information section in navigation.
-	SectionWeight = 100
-)
